commerce_api/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
Flask-Marshmallow>=1.2.0
marshmallow>=3.21.0
marshmallow-sqlalchemy>=1.0.0
Flask-Limiter>=3.5.0
pytest>=8.0.0
python-dotenv>=1.0.0
